:art: Split Yugioh forms in respective domains
