+++ source/YugiAcademy/YuGiOh/booster_packs/forms/search_booster_pack_form.py
@@ -1,0 +1,8 @@
+from django import forms
+
+
+class SearchBoosterPackForm(forms.Form):
+    name = forms.CharField(max_length=50)
+
+    def __str__(self):
+        return 'Form for searching booster packs through their name'
