+++ source/YugiAcademy/YuGiOh/cards/forms/__init__.py
@@ -1,0 +1,1 @@
+from .spell_card_form import SpellCardForm
